requirements.txt
Flask==2.3.2
gunicorn==20.1.0
python-dotenv==0.19.2
python-dotenv
requests==2.31.0